Break out errors and loading into separate hocs
