+++ hocs/withLoadingAndErrors.js
@@ -1,34 +1,8 @@
-import React, { Component } from 'react'
-import { compact, some, map, pick } from 'lodash'
+import { compose } from 'react-apollo'
 
-import ErrorScreen from '../components/ErrorScreen'
-import LoadingScreen from '../components/LoadingScreen'
+import withLoading from './withLoading'
+import withErrors from './withErrors'
 
-const DEFAULT_KEYS = ['data']
+export default (WrappedComponent, options = {}) =>
+  compose(withLoading, withErrors)(WrappedComponent, options)
 
-export default (WrappedComponent, { errorMessage, dataKeys = DEFAULT_KEYS } = {}) => {
-  class WithLoadingAndErrors extends Component {
-    render() {
-      const datums = pick(this.props, dataKeys)
-      const isLoading = some(map(datums, 'loading'))
-      const errors = compact(map(datums, 'error'))
-
-      if (isLoading) {
-        return <LoadingScreen />
-      }
-
-      if (errors.length > 0) {
-        return (
-          <ErrorScreen
-            message={errorMessage}
-            errors={errors}
-          />
-        )
-      }
-
-      return <WrappedComponent {...this.props} />
-    }
-  }
-
-  return WithLoadingAndErrors
-}
